refactor lntest as it was getting large
