+++ libnode2/src/lib.rs
@@ -4,3 +4,6 @@ pub mod logging;
 pub mod pki;
 pub mod vsconn;
 pub mod vss;
+
+#[cfg(feature = "build-lnt")]
+pub mod cli;
